triapp/requirements.txt
boto3
azure-storage-blob
google-cloud-pubsub
